# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Vercel OpenAPI Spec",
        "title": "Vercel OpenAPI Spec",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://vercel-echo-functions.vercel.app",
    "basePath": "/",
    "paths": {
        "/device/{platform}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get list of devices",
                "tags": [
                    "Devices"
                ],
                "summary": "Get list of devices",
                "parameters": [
                    {
                        "enum": [
                            "1",
                            "2"
                        ],
                        "type": "string",
                        "description": "platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Device"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Device": {
            "type": "object",
            "required": [
                "location",
                "name"
            ],
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}